Configure provider error retry backoff
Add finite and explicit unlimited retry policy for transient provider outages, cap exponential backoff at fifteen minutes, and preserve monotonic timer generations. Wire live configuration through schema v81 with migration, telemetry, cancellation coverage, and operator documentation.

+++ crates/mezzanine/src/host/async_runtime/actor/events.rs
@@ -1100,7 +1100,7 @@ impl AsyncRuntimeSessionActor {
                         )?
                         && self
                             .service
-                            .queue_agent_provider_retry_task(&turn_id, attempt)?
+                            .queue_agent_provider_retry_task(&turn_id, u64::from(attempt))?
                     {
                         let mut side_effects =
                             self.render_side_effects(RenderInvalidationReason::FullRedraw);
